requirements.txt
streamlit
opencv-python
numpy
pandas
Pillow
python-dotenv
google-generativeai
pyngrok
easyocr
symspellpy